+++ src/reducers/socketReducer.test.js
@@ -2,27 +2,33 @@ import socketReducer from './socketReducer'
 import * as types from '../actions/ActionTypes';
 
 describe('SocketReducer Reducer', () => {
-  const initialState = false
+  const initialState = {
+    connected: false,
+    loading: false
+  }
 
   it('should return the initial state', () => {
     expect(socketReducer(undefined, {})).toEqual(initialState)
   })
   it('should handle SOCKET_WRITE', () => {
+    const newState = { connected: false, loading: true }
     const action = { type: types.SOCKET_WRITE }
-    expect(socketReducer(undefined, action)).toEqual(true)
-    expect(socketReducer(true, action)).toEqual(true)
-    expect(socketReducer(false, action)).toEqual(true)
+    expect(socketReducer(undefined, action)).toEqual(newState)
   })
-  it('should handle SOCKET_WRITE_SUCCESS', () => {
+  it('should handle LOAD_FILES_SUCCESS, UPDATE_FILE_SUCCESS, DELETE_FILE_SUCCESS, ADD_FILE_SUCCESS and SOCKET_WRITE_FAILURE', () => {
+    const newState = { connected: false, loading: false }
+    const actions = [
+      { type: types.LOAD_FILES_SUCCESS },
+      { type: types.UPDATE_FILE_SUCCESS },
+      { type: types.DELETE_FILE_SUCCESS },
+      { type: types.ADD_FILE_SUCCESS },
+      { type: types.SOCKET_WRITE_FAILURE }
+    ]
+    actions.forEach((action) => expect(socketReducer(undefined, action)).toEqual(newState))
+  })
+  it('should handle SOCKET_CONNECT_SUCCESS', () => {
+    const newState = { connected: false, loading: false }
     const action = { type: types.SOCKET_WRITE_SUCCESS }
-    expect(socketReducer(undefined, action)).toEqual(false)
-    expect(socketReducer(true, action)).toEqual(false)
-    expect(socketReducer(false, action)).toEqual(false)
-  })
-  it('should handle SOCKET_WRITE_FAILURE', () => {
-    const action = { type: types.SOCKET_WRITE_FAILURE }
-    expect(socketReducer(undefined, action)).toEqual(false)
-    expect(socketReducer(true, action)).toEqual(false)
-    expect(socketReducer(false, action)).toEqual(false)
+    expect(socketReducer(undefined, action)).toEqual(newState)
   })
 })